requirements.txt
pylast
winsdk
python-dotenv
bleak
pycaw
comtypes
customtkinter
Pillow
requests